Move CI-visibility classes from internal-api to agent-ci-visibility module
