+++ dd-java-agent/agent-ci-visibility/build.gradle
@@ -15,6 +15,8 @@ excludedClassesCoverage += [
   "datadog.trace.civisibility.DDTestSessionImpl",
   "datadog.trace.civisibility.DDTestSuiteImpl",
   "datadog.trace.civisibility.DDTestImpl",
+  "datadog.trace.civisibility.ci.CIInfo",
+  "datadog.trace.civisibility.ci.CIInfo.Builder",
   "datadog.trace.civisibility.context.AbstractTestContext",
   "datadog.trace.civisibility.context.EmptyTestContext",
   "datadog.trace.civisibility.context.ParentProcessTestContext",
@@ -27,6 +29,7 @@ excludedClassesCoverage += [
   "datadog.trace.civisibility.events.TestModuleDescriptor",
   "datadog.trace.civisibility.events.TestSuiteDescriptor",
   "datadog.trace.civisibility.git.GitObject",
+  "datadog.trace.civisibility.source.MethodLinesResolver.MethodLines",
 ]
 
 dependencies {
