+++ dd-java-agent/agent-ci-visibility/src/main/java/datadog/trace/civisibility/git/CIProviderGitInfoBuilder.java
@@ -1,8 +1,8 @@
 package datadog.trace.civisibility.git;
 
-import datadog.trace.api.civisibility.ci.CIProviderInfo;
 import datadog.trace.api.git.GitInfo;
 import datadog.trace.api.git.GitInfoBuilder;
+import datadog.trace.civisibility.ci.CIProviderInfo;
 import datadog.trace.civisibility.ci.CIProviderInfoFactory;
 import java.nio.file.Path;
 import java.nio.file.Paths;
