+++ dd-java-agent/instrumentation/junit-5.3/src/test/groovy/SpockTest.groovy
@@ -1,5 +1,5 @@
 import datadog.trace.agent.test.asserts.ListWriterAssert
-import datadog.trace.agent.test.base.CiVisibilityTest
+import datadog.trace.civisibility.CiVisibilityTest
 import datadog.trace.api.DisableTestTrace
 import datadog.trace.api.civisibility.CIConstants
 import datadog.trace.bootstrap.instrumentation.api.Tags
